pyproject.toml
[project]
name = "career-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.1.1",
    "openai-agents",
    "python-dotenv>=1.1.1",
]

[tool.uv.sources]
openai-agents = { git = "https://github.com/openai/openai-agents-python" }
